image insert,update,and deleted today are completed

+++ database/migrations/2024_07_24_013136_create_project_table.php
@@ -14,7 +14,7 @@ class CreateProjectTable extends Migration
     public function up()
     {
         Schema::create('project', function (Blueprint $table) {
-            $table->integer('Id');
+            $table->bigIncrements('Id');
             $table->string('Name',30);
             $table->string('Address',60);
             $table->string('Email_verifed_at',50);
